Drastically simplify
This is the reason you need to sleep on code.  There is a better way
than copy-n-pasting the entire autodiscover mechanism.  Instead, we rely
Django's auto-discover, then merge any of our registered sites into what
was registered with the default admin.

This is still less than ideal, but such is life with globals.